Simplify Tests and libraries Simplify Tests and libraries Simplify Tests and libraries
- Rename script_helpers to helpers to make more generic for future unit testing
- Test kitchen: Use new slugs for platforms
- Make the dokken suites the same as vagrant by incuding it when needed
in the test recipes
- Update README to note that we always require a target ruby version for
the commands. This makes it unambiguous as to which version we're
installing it to
