--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "foodcritic/sc"]
-	path = foodcritic/sc
-	url = https://github.com/sous-chefs/sc-foodcritic-rules.git
