+++ .kitchen.yml
@@ -11,28 +11,23 @@ verifier:
   name: inspec
 
 platforms:
-  - name: centos-6.9
-  - name: centos-7.3
+  - name: centos-6
+  - name: centos-7
   - name: fedora-25
-  - name: debian-8.8
-  - name: debian-9.0
+  - name: debian-8
+  - name: debian-9
   - name: ubuntu-14.04
   - name: ubuntu-16.04
-  - name: ubuntu-16.04-chef-12
-    provisioner:
-      require_chef_omnibus: 12
-    driver_config:
-      box: bento/ubuntu-16.04
 
 suites:
 - name: gem
   run_list: recipe[test::gem]
-  includes: [centos-7.3]
+  includes: [centos-7]
 - name: global
   run_list: recipe[test::global]
 - name: system_install
   run_list: recipe[test::system_install]
-  includes: [centos-7.3]
+  includes: [centos-7]
 - name: user_install
   run_list: recipe[test::user_install]
-  includes: [centos-7.3]
+  includes: [centos-7]
