+++ .travis.yml
@@ -22,7 +22,6 @@ env:
     - INSTANCE=global-debian-8
     - INSTANCE=global-debian-9
     - INSTANCE=global-centos-7
-    - INSTANCE=global-centos-7 CHEF_VERSION=12.9.41
     - INSTANCE=global-fedora-latest
     - INSTANCE=global-ubuntu-1604
     - INSTANCE=global-ubuntu-1804
