+++ resources/rehash.rb
@@ -33,5 +33,5 @@ action :run do
 end
 
 action_class do
-  include Chef::Rbenv::ScriptHelpers
+  include Chef::Rbenv::Helpers
 end
