+++ resources/script.rb
@@ -51,5 +51,5 @@ action :run do
 end
 
 action_class do
-  include Chef::Rbenv::ScriptHelpers
+  include Chef::Rbenv::Helpers
 end
