+++ test/fixtures/cookbooks/test/recipes/gem.rb
@@ -1,12 +1,12 @@
-rbenv_system_install 'system'
-rbenv_user_install 'vagrant'
+# Make sure that Vagarant user is on the box for dokken
+include_recipe 'test::dokken'
 
+# System Install
+rbenv_system_install 'system'
+# Install several rubies to /opt/rubies
 rbenv_ruby '2.4.1'
 rbenv_ruby '2.3.1'
-rbenv_ruby '2.3.1' do
-  user 'vagrant'
-end
-
+# Set System global version
 rbenv_global '2.4.1'
 
 rbenv_gem 'mail' do
@@ -16,9 +16,29 @@ rbenv_gem 'mail' do
 end
 
 rbenv_gem 'mail' do
-  version '2.6.6'
-  rbenv_version '2.3.1'
+  version '2.6.5'
+  options '--no-rdoc --no-ri'
+  rbenv_version '2.4.1'
+end
+
+rbenv_gem 'mail' do
+  version '2.6.5'
+  rbenv_version '2.4.1'
   action :remove
+end
+
+
+# User Install
+rbenv_user_install 'vagrant'
+
+# Install a Ruby to a user directory (~/.rubies)
+rbenv_ruby '2.3.1' do
+  user 'vagrant'
+end
+
+# Set the vagrant global version
+rbenv_global '2.3.1' do
+  user 'vagrant'
 end
 
 rbenv_gem 'bundler' do
