+++ test/fixtures/cookbooks/test/recipes/global.rb
@@ -8,5 +8,8 @@ rbenv_ruby version do
   verbose true
 end
 
+# Make sure that Vagarant user is on the box for dokken
+include_recipe 'test::dokken'
+
 # Set that Ruby as the global Ruby
 rbenv_global version
